+++ gatsby-config.js
@@ -2,6 +2,9 @@ module.exports = {
   siteMetadata: {
     title: `Gatsby Default Starter`,
   },
-  plugins: [`gatsby-plugin-react-helmet`],
+  plugins: [
+    `gatsby-plugin-react-helmet`,
+    `gatsby-plugin-stylus`,
+  ],
   pathPrefix: `/geoWatcher`,
 }
