+++ src/components/Header.js
@@ -1,0 +1,32 @@
+import React from 'react'
+import PropTypes from 'prop-types'
+import Link from 'gatsby-link'
+import Helmet from 'react-helmet'
+import '../styles/header.styl';
+
+const Header = () => (
+    <div className="headerContainer">
+      <div
+        style={{
+          margin: '0 auto',
+          maxWidth: 960,
+          padding: '1.45rem 1.0875rem',
+        }}
+      >
+        <h1 style={{ margin: 0 }}>
+          <Link
+            to="/"
+            style={{
+              color: 'white',
+              textDecoration: 'none',
+            }}
+          >
+            Geo Watcher
+          </Link>
+        </h1>
+      </div>
+    </div>
+  )
+
+export default Header
+
