+++ src/pages/ngarkat.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import Link from 'gatsby-link'
+
+const Ngarkat = () => (
+  <div>
+    <iframe width="750" height="633" src="http://earthengine.google.org/timelapse/player?c=http%3A%2F%2Fearthengine.google.org%2Ftimelapse%2Fdata#tour=DhBDyBAmDkQiOAxoZ1984_DyBBmDkQiOAxoZ_DyBCmDkQiOAxoZ_DyBDmDkQiOAxoZ_DyBEmDkQiOAxoZ_DyBFmDkQiOAxoZ_DyBGmDkQiOAxoZ_DyBHmDkQiOAxoZ_DyBImDkQiOAxoZ_DyBJmDkQiOAxoZ_DyBKmDkQiOAxoZ_DyBLmDkQiOAxoZ_DyBMmDkQiOAxoZ_DyBNmDkQiOAxoZ_DyBOmDkQiOAxoZ_DyBPmDkQiOAxoZ_DyBQmDkQiOAxoZ_DyBRmDkQiOAxoZ_DyBSmDkQiOAxoZ_DyBTmDkQiOAxoZ_DyBUmDkQiOAxoZ_DyBVmDkQiOAxoZ_DyBWmDkQiOAxoZ_DyBXmDkQiOAxoZ_DyBYmDkQiOAxoZ_DyBZmDkQiOAxoZ_DyBamDkQiOAxoZ_DyBbmDkQiOAxoZ_DyBcmDkQiOAxoZ_DyBdmDkQiOAxoZ_DyBemDkQiOAxoZ_DyBfmDkQiOAxoZ_BkDgBmDkQiOAxoZ_Untitled_B" frameborder="0"></iframe>
+    <h1>Hi from the second page</h1>
+    <p>Welcome to page 2</p>
+    <Link to="/">Go back to the homepage</Link>
+  </div>
+)
+
+export default Ngarkat
