+++ src/styles/header.styl
@@ -1,0 +1,2 @@
+.headerContainer
+  background: green;
